wait for event made common to both schedulers
